// src/app/models/IZPL_EtiquetaFija.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import { IZPL_TipoEtiquetas } from "./IZPL_TipoEtiquetas";

export interface IZPL_EtiquetaFija {
    id: number;
    modelo: string;
    zpl: string;
    tipoUnidad: string;
    zpL_TipoEtiquetasId: number;
    modelosId: number;
    zpL_TipoEtiquetas: IZPL_TipoEtiquetas;
    preFijo: string;
  }
                                                    